# crypto, refactor: add new KeyPair class