@theuni 2024-05-07
`ComputeTapTweak` is more like what I had in mind, yes.