@josibake 2024-05-08
Updated to use a utility function (instead of a method on `CKey`) per the @theuni 's feedback