@josibake 2024-05-10
Updated to use the new `KeyPair` wrapper class (h/t @theuni).