@josibake 2024-07-21
Update https://github.com/bitcoin/bitcoin/commit/5d9a6cf6286f9a7f93527ea76b910537d709a860 ->  https://github.com/bitcoin/bitcoin/commit/49057fc4f1e73a14f673c934573d727ae0229779 ([apply-taptweak-method-02](https://github.com/josibake/bitcoin/tree/apply-taptweak-method-02) -> [apply-taptweak-method-03](https://github.com/josibake/bitcoin/tree/apply-taptweak-method-03) ([compare](https://github.com/josibake/bitcoin/compare/apply-taptweak-method-02..josibake:apply-taptweak-method-03))

* Use `CKey::GetPubKey()` instead of extracting and serializing the public key from the keypair object
* Save `reinterpret_cast` as a variable to simplify diff/code (h/t @paplorinc )

Overall, I think this simplifies things quite a bit. If you look at the internals of `CKey::GetPubKey()`, there are already `asserts` for intermediate libsecp256k1 function calls which make it clear these calls (`_serialize`, for example) are always expected to succeed when using a valid secret key.