@josibake 2024-07-21
Update https://github.com/bitcoin/bitcoin/commit/49057fc4f1e73a14f673c934573d727ae0229779 - > https://github.com/bitcoin/bitcoin/commit/b8b3a9f18670ec7bf246a57950cdae7e034a264d([apply-taptweak-method-03](https://github.com/josibake/bitcoin/tree/apply-taptweak-method-03) -> [apply-taptweak-method-04](https://github.com/josibake/bitcoin/tree/apply-taptweak-method-04) ([compare](https://github.com/josibake/bitcoin/compare/apply-taptweak-method-03..josibake:apply-taptweak-method-04))

* Renamed `ret` to `success` in KeyPair constructor (h/t @paplorinc)
* Removed unnecessary `KeyPair()` call
* Added key tweak smoke test (h/t @paplorinc )