@paplorinc 2024-07-22
CI failure seems unrelated, please restart or rebase so we can ACK - and don't worry about our reacks :)

> Update https://github.com/bitcoin/bitcoin/commit/b8b3a9f18670ec7bf246a57950cdae7e034a264d -> https://github.com/bitcoin/bitcoin/commit/56267cf84281789186035f09898cce3d749748ec ([apply-taptweak-method-04](https://github.com/josibake/bitcoin/tree/apply-taptweak-method-04) -> [apply-taptweak-method-05](https://github.com/josibake/bitcoin/tree/apply-taptweak-method-05) ([compare](https://github.com/josibake/bitcoin/compare/apply-taptweak-method-04..josibake:apply-taptweak-method-05))

I usually review the changes first via:
<img src="https://github.com/user-attachments/assets/4356aa17-72a2-4198-af6c-8082e102edac">